functions/embed_and_index/requirements.txt
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.14
jmespath==1.0.1
numpy==1.26.4
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
